django_piston/requirements.txt
django>=1.3
django-piston
